+++ src/web/streams/mod.rs
@@ -1,10 +1,12 @@
 pub mod readable;
 pub mod strategy;
+pub mod writable;
 
 use boa_engine::{Context, JsResult};
 
 pub fn register_globals(context: &mut Context) -> JsResult<()> {
     strategy::register_globals(context)?;
     readable::register_globals(context)?;
+    writable::register_globals(context)?;
     Ok(())
 }